Move the assembly walk into GatewayCore
It lived inside `RenderService`, in the app target, which no command-line
tool can import — so no `gfcheck` suite could ever run it. Two faults
reached you through that gap and neither was subtle: `@pan` applied to a
whole session instead of the segment that declared it, leaving every word
after Headphone Orientation in one ear; and the resonant tuning never
sounding at all. They were not hard to catch. There was simply nowhere to
catch them from.

Nothing about the walk changed in the move. `RenderService` keeps what
belongs to it — scanning, resolving, rendering missing takes, choosing
the folder, writing the files, saying so on screen — and is 199 lines
lighter. `SessionAssembly` decides only what the tape contains.

`gfcheck` now has an assembly suite that builds its own takes, so it runs
on a checkout with no rendered audio. Putting either shipped fault back
fails it: the session-wide pan on one check, the silent tuning on three.

And the comparison this unlocks: `gfcorpus assembly-fixture` walks the
same five steps over takes built from the same arithmetic the TypeScript
check builds them from, at two pause scales. Neither side reads the
other's audio; neither reads a library. They now agree on the narration
byte for byte and on every field of the written manifest. Diverging the
pace rule or dropping a piece's pan names the exact field and value.

Reaching byte-for-byte cost one real fix. Swift's samples are `Float`, so
`s * 32767` is a single-precision multiply that rounds before `Int16(...)`
truncates, and JavaScript did it in double — fourteen samples in every
48,000 of a plain sine came out one integer apart. Inaudible, and still
the difference between the two builds producing the same tape and merely
producing tapes that sound alike. Only the first can be checked.

+++ cross-platform/src/core/audioIO.ts
@@ -109,7 +109,21 @@ function header(channels: number, frames: number, rate: number): Buffer {
   return b;
 }
 
-const toPCM = (v: number): number => Math.trunc(Math.max(-1, Math.min(1, v)) * 32767);
+/**
+ * A sample as the 16-bit integer a wav carries.
+ *
+ * **The scaling happens in single precision**, because Swift's `AudioIO` does:
+ * its samples are `Float`, so `s * 32767` is a `Float` multiply that rounds to
+ * nearest before `Int16(...)` truncates. JavaScript would do the same multiply
+ * in double and truncate a slightly different number — for 14 samples in every
+ * 48 000 of a plain sine, one integer lower.
+ *
+ * Inaudible, and still worth matching: it is the difference between the two
+ * builds producing the same tape and merely producing tapes that sound alike,
+ * and only the first can be checked.
+ */
+const toPCM = (v: number): number =>
+  Math.trunc(Math.fround(Math.max(-1, Math.min(1, v)) * 32767));
 
 export function writeWav(samples: Float32Array, path: string, rate = sampleRate): void {
   const body = Buffer.alloc(samples.length * 2);
